Fix lib item spec.libraryName conversion
This patch updates the conversion logic for [Cluster]ContentLibraryItem
objects and their field spec.libraryName. If the field is non-empty in
v1a2, it should be set from the value status.contentLibraryRef.name
from the v1a1 object. However, because that logic was *after* the code
that checked for the marshaled restore data, if said data was no present,
the library name conversion would not run.

This patch moves the library name initialization to a deferred function
that always runs as part of spoke->hub conversion.

A test has also been added to validate this.

The reason this was not caught sooner is because on real testbeds that
used to have v1a1 objects, their status.contentLibraryRef.name fields
were already set. Thus when converting from v1a2->v1a1, the only reason
it looked like those fields were non-empty is not because of conversion,
but because they were originally non-empty.